wb-infographic/worker/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
rembg>=2.0.50
Pillow>=10.0.0
